Fix `AnyDependency::Emplace<AnyDependency>()`.
Returning a reference to the newly constructed object is incompatible with the
optimization of replacing wrapping `AnyDependency` with adopting it (i.e. moving
its internals), because after adoption there is no object of the right type to
return a reference to. The existing code returns a reference which does not
actually point to the object of the type it promises.

The optimization of avoiding wrapping is useful when `AnyDependency` is
constructed from or assigned from a similar but different `AnyDependency`.
But explicitly providing constructor arguments (whether in `Emplace()` or by
providing a tuple of constructor arguments) is not useful if the object to be
constructed is another `AnyDependency` (in practice the constructor arguments
can be adapted to construct the outer `AnyDependency` instead), and this case
is supported only for completeness (disallowing that would be an artificial
limitation).

Hence the fix is simple: when constructing the `Manager` from a tuple of
constructor arguments (this is used also by `Emplace()`), stop treating another
`AnyDependency` specially. This implies that it will be wrapped instead of
adopting its internals even if it would fit in the target type.

PiperOrigin-RevId: 522264635