Reconsider `ReadAll()` to `absl::string_view`: remove the variant which specifies processing as a function which is called also if reading failed (which includes exceeding `max_length`).
It is too weird, unintuitive, and it requires a special `StatusAnd<T>` type to
express the result.

Instead, the caller can manage ownership of the `Reader` explicitly (perhaps
just skip calling `Close()` it if neither the status of closing nor possible
side effects of that are important). Or read to `std::string`.

Also, remove support for a function returning a reference, because there is no
natural way to wrap a reference in `absl::StatusOr`. Such a function can be
changed to return either a pointer or a value.

PiperOrigin-RevId: 523499266