Change `{Linear,Chunked}SortedStringSet::Iterator` API to standard input iterators, with `operator*` returning an `absl::string_view` pointing to a string owned by the iterator.
Use a pair of `const char*` instead of `StringReader` in
`LinearSortedStringSet::Iterator` to make the iterator copyable and to reduce
overhead.

Use `CompactString` instead of `std::string` in
`LinearSortedStringSet::Iterator` to make it lighter. Avoid repeated conversion
of it to `absl::string_view`.

Avoid `absl::string_view::substr()` when the index is known to be within bounds.
This avoids a bounds check leading to `abort()`.

PiperOrigin-RevId: 535573891