Use `ResizeStringAmortized()` in `SerializeToString()` to ensure that repeated calls with the same string but slowly growing sizes amortize the cost of reallocatiosn.
Add `SerializeToCompactString()`.

Use `ArrayWriter` instead of `StringWriter` in `SerializeToString()` when
`options.deterministic()` is not the default and a `Writer` must be used.
The exact size is known beforehand so this is simpler.

Move the assertion that `ByteSizeLong()` did not change before checking for
serialization failures, so that concurrent modification of the message results
in a more specific failure than `ArrayWriter` overflowing.

PiperOrigin-RevId: 540224903