Redo the `StringResizableTraits` change with a fix.
Use `std::string().capacity()` instead of `std::basic_string<_>.capacity()` for
the minimal capacity. The latter cannot be used (not all `std::basic_string`
instantiations are default-constructible), in practice they should be the same
(at least in libc++ and libstdc++ the minimal capacity does not depend on the
allocator as long as the `pointer` type is a real pointer), and even if they are
not the same, this is a matter of performance tuning, not correctness.

PiperOrigin-RevId: 544015259