Add `Dec()` and `Hex()` for stringifying integer values as decimal or hexadecimal, padding with zeros to at least the given width.
Differences wrt. `absl::Dec()` and `absl::Hex()`:

* Only zero padding is supported, no space padding. (Padding with other
  characters will be provided separately, with a different API.)
* Padding width can be a variable rather than a constant up to 20.
* They are faster when used with a `Writer`, thanks to using `Writer::Push()`
  and formatting the value directly in the destination rather than formatting in
  a temporary buffer and copying.
* They support `absl::{,u}int128`.
* Their type is templatized wrt. the value rather than fixed to contain
  `uint64_t` and sign. This enables tuning the argument of `Push()`
  (asking for the amount of space appropriate for the type) and supporting
  `absl::{,u}int128`.

Inline the outer part of `Writer::Write()` for integers.

Extract common parts of writing non-negative and negative values of a signed
number to reduce code size, at the cost of calling `Push()` with space for a
sign also in the non-negative case.

This adds `riegeli/text` subdirectory. Contents of `riegeli/lines` will be moved
to it.

PiperOrigin-RevId: 549312117