Fix build after https://github.com/llvm/llvm-project/commit/e30a148b098d462d0267c479cd9e4783363c2761.
`std::char_traits` is not guaranteed to be defined for `signed char` and
`unsigned char`, and libc++ is removing the base template.

It is simpler to avoid relying on `std::char_traits<T>::int_type` and just
hardcode integer types providing a numeric `operator<<` covering values of each
character type. Character types are listed explicitly anyway.

PiperOrigin-RevId: 555862881