Add comments that buffer pointers may be changed by non-const member functions Only `{move,set}_cursor()` are guaranteed to move only `cursor()` but not `start()` nor `limit()`.
Remove the comment that `{,Backward}Writer::set_cursor()` can be used to seek
within the buffer. This is in general not supported if the `Writer` does not
support random access (and `BackwardWriter` never does). Instead, add comments
that memory after the address to which `cursor()` is eventually moved must not
be clobbered. This is in practice unsafe when the current position is not at
the end, because some `Writer` implementations expose pointers to an in-memory
destination; for consistency disallow it for `BackwardWriter` too, should it
optionally support random access some day.

PiperOrigin-RevId: 558144584