Optimize `LinearSortedStringSet::contains()` by introducing and using `LinearSortedStringSet::SplitElementIterator` which represents each element as the concatenation of two `absl::string_view` values: prefix and suffix. This is more efficient than a single `absl::string_view` but less convenient.
Remove `{Linear,Chunked}SortedStringSet::Iterator::Next()` which was a variant
of `operator++` which returned a shared length. The shared length was used only
in `LinearSortedStringSet::contains()`, which now reles on the promise that each
`SplitElement::prefix()` is shared with the previous element.

This speeds up `{Linear,Chunked}SortedStringSet::contains()` by about 10%.

PiperOrigin-RevId: 565665923