Remove the just added `RecordWriter::PadToBlockBoundary()` function. This will be done differently.
A problem with `PadToBlockBoundary()` is that if the file was just created, it
will have file signature and possibly file metadata written, so padding will be
written after them even though this is not needed. The `pad_to_block_boundary`
option does not have this problem because it checks whether padding is needed
before writing the signature and metadata.

PiperOrigin-RevId: 566394394