Let `Digesting{Reader,Writer}` and `DigestFrom{,Reader}()` require the `DigesterType` to derive from `DigesterBase` and have a `Digest()` member function instead of relying on static matching of function names.
Embrace the `Dependency` mechanism for getting a digester out of a parameter
which provides and possiblt owns it, replacing the previous ad hoc mechanism to
follow `operator*` if needed. This changes the behavior regarding `Close()`:
it is called only when the digester is owned, i.e. not for raw pointers.
`ClosingPtr(&digester)` can be used to let `Digesting{Reader,Writer}` call
`Close()` without moving the object.

Change how `Digesting{Reader,Writer}` pass constructor parameters to the
digester. Instead of forwarding all remaining constructor parameters of
`Digesting{Reader,Writer}`, use the same convention as most Riegeli classes:
pass an already constructed digester or a tuple of its constructor arguments.
By default the digester is default-constructed.

Change how the digester is specified for `DigestFrom{,Reader}()`: they only
accept an already constructed digester, removing the possibility to explicitly
specify the digester type in a template parameter and forward any number of its
constructor parameters starting from the second parameter of
`DigestFrom{,Reader}()`. The new convention is as functional and as efficient,
because the functions do not move from the parameter if it is a temporary,
because they do not need the object to be valid after they return.

Remove `DigestFromTuple()`. Instead, make `DigestFrom()` variadic. This is now
unambiguous because the digester is always a single last parameter.

Fix `DigestFrom(const char*)` which did not compile because an internal function
did not explicitly specify a template parameter.

Optimize `DigestFrom()` by skipping constructing the `DigestingWriter` for all
types directly supported by `DigesterBase`, also with multiple arguments.

PiperOrigin-RevId: 571079392