Redirect non-inlined friend functions through private member functions, so that they have a definition in the class.
It is said that in such case the friend function is available only through ADL
without polluting the namespace.

Experiments indicate that this is true in MSVC, while in Clang and GCC a friend
function is available only through ADL no matter whether it is defined or
declared in the class. It does not hurt to follow this convention.

PiperOrigin-RevId: 578854937