Rename `Wrapped{Reader,Writer,BackwardWriter}` to `Wrapping{Reader,Writer,BackwardWriter}`.
Rationale: the general naming pattern of Riegeli classes uses -ing for
standalone classes to be instantiated (e.g. `LimitingReader`), and -ed for
abstract classes providing implementation strategies for derived classes
(e.g. `BufferedReader`). This was an exception.

Temporarily leave `Wrapped{Reader,Writer}` wrappers for migration (not aliases
because of CTAD).

PiperOrigin-RevId: 589056620