Fix building with gcc: make the partial specialization of `WrappingDigester` obviously more specialized than the primary template, rather than letting it happen to be more specialized for types which match SFINAE constraints.
Clang allows either, but gcc makes this a warning or error:
`partial specialization […] is not more specialized than [-fpermissive] […]`.

The issue is with the type of `digest_converter`, which should be
`DigestType (*)(DigestTypeOf<BaseDigesterType>)`. The spelling
`DigestType (*)(DigestType)` previously used in the specialization is the same
only due to `std::is_base_of<Digester<DigestType>, BaseDigesterType>`.

PiperOrigin-RevId: 590181781