When constructing `DependencyBase`, `StableDependency`, or `AnyDependency` from `Manager&&`, check for *implicit* copy/move constructibility of `std::decay_t<Manager>` rather than also explicit one.
In some build configurations (notably on Android), and also on a regular Linux
buidl with some planned changes, a check which allows explicit constructibility
interferes with moving of related types.

For example, to decide whether `BrotliReader<AnyDependency<Reader*>>` is movable
(e.g. to generate a defaulted move constructor of a class containing it),
the compiler considers not only the move constructor of
`BrotliReader<AnyDependency<Reader*>>`, but also the constructor of
`BrotliReader<Src>` from `Src&&`, i.e. from `AnyDependency<Reader*>&&`.

This is so even though the other constructor is explicit, and even though it is
not intended to be used here: we want to move the `BrotliReader`, not to wrap it
in another `BrotliReader`, and the move constructor would be a better match, not
requiring a conversion of the parameter.

Considering the other constructor requires converting
`BrotliReader<AnyDependency<Reader*>>&&` to `AnyDependency<Reader*>`.

This conversion checks whether `BrotliReader<AnyDependency<Reader*>>` is move
constructible.

This is what we are still checking! A self-dependent template instantiation
breaks the build in a non-SFINAE-friendly way. Symptom: `std::conjunction<T> : T
{}` fails for `T` being `std::is_constructible<...>` when this is an incomplete
type.

Also, converting `BrotliReader<AnyDependency<Reader*>>&&`
to `AnyDependency<Reader*>` requires checking whether
`IsValidDependency<Reader*, BrotliReader<AnyDependency<Reader*>>>`.
If this instantiates `DependencyBase<BrotliReader<AnyDependency<Reader*>>>`,
including its constructor from `BrotliReader<AnyDependency<Reader*>>&&`, this
constructor should not depend on whether `BrotliReader<AnyDependency<Reader*>>`
is move constructible.

The fix is to use `std::is_convertible<const T&, T>` instead of
`std::is_copy_constructible<T>`, and `std::is_convertible<T&&, T>` instead of
`std::is_move_constructible<T>` when checking whether an arbitrary `Manager` is
copy/move constructible. Explicit copy/move constructors are dubious anwyay.

The fixed version does not have the same problem because the other constructor
of `BrotliReader` etc., being explicit, is rejected before trying to convert its
parameter.

PiperOrigin-RevId: 599478846