In `describe_riegeli_file` tool, default `--show_records_metadata` to `false`. Metadata can be large, let’s show the contents only when requested.
Intersperse failures with the rest of the report, so that they can be more
easily attributed to chunks. The convention of formatting failures like comments
with `#` keeps the overall feel of the report as a text proto.

Improve annotations of statuses so that they look more like how `RecordReader`
annotates them.

PiperOrigin-RevId: 604910820