Let `StatusOrMaker` support references by dereferencing them.
This was previously left unsupported because of an unclear intent: whether to
dereference them, or to convert to a pointer so that they can be put in
`absl::StatusOr`, or wait until some kind of `StatusOrRef<T>` emerges in future.

But it is risky to store the address, especially if this behavior is inferred
from the function returning a reference, because a `StatusOr` often outlives the
full expression containing the function call, while the reference often points
to a temporary created in that expression. Storing the value is a safer deafult.

If storing the address is intended, the function can be changed to return a
pointer instead.

Also, get rid of `read_all_internal::StringViewCallResult` to reduce
abstraction layers. Use `StatusOrMaker` directly instead. Leave a different
`read_all_internal::StringViewCallResult` which is an alias to the result of
applying the function to `absl::string_view`.

PiperOrigin-RevId: 605587807