Change various functions which read the whole source to call `src.VerifyEnd()` only if there were no other failures so far.
If parsing failed, or if they are writing to a `Writer` and writing failed, then
the source is not expected to end, and making it failed can be misleading.

This does not make a difference if any failures must have been caused by the
`Reader` and it is necessarily already failed, because `VerifyEnd()` would keep
the existing failure, but this pattern is applied everywhere for consistency.

PiperOrigin-RevId: 605587886