Rename terminology from `Dependency<Ptr, Manager>` to `Dependency<Handle, Manager>`.
There will be more cases where `Handle` is not a pointer, as well as
related pointers which are not `Ptr`, so `Ptr` can be misleading.

Rename template parameters to use fewer one-letter names.

Rename internal `DependencyMaybeRef` to `DependencyCore`. It will acquire more
duties so the old name will be insufficient.

PiperOrigin-RevId: 606180269