Improve performance of block allocation in `Cord{,Backward}Writer` by avoiding the case when `absl::CordBuffer` allocates less than required.
`absl::CordBuffer::CreateWithCustomLimit()` treats the requested capacity not as
a requirement but as a hint. It promises to return a block with at least that
capacity only up to `absl::CordBuffer::kDefaultLimit` (4K-13). Above that it
rounds the capacity either up or down to a power of 2, excluding `kFlatOverhead`
(13), depending on how much memory would be wasted.

`CordWriter` sometimes requires a particular length, not only a hint. If
`absl::CordBuffer` does not satisfy the required length, the `absl::CordBuffer`
is ignored and `CordWriter` reverts to using `Buffer` instead. This may cause
repeated wasted allocations, especially if the same sizes are used every time.

Hence we ensure that the requested capacity covers at least the required length
rounded up in the appropriate way, to ensure that if `absl::CordBuffer` is asked
to allocate then it will allocate enough. Since the exact computation by
`absl::CordBuffer` is not guaranteed, `CordWriter` still reverts to `Buffer`
if needed. But now this should no longer happen.

This reduces the number of allocations by the ignored `absl::CordBuffer`
creations. Also, allocating an `absl::CordBuffer` and appending it to an
`absl::Cord` needs fewer allocations than with a `Buffer`. OTOH unlucky
configured block sizes and required buffer sizes cause more memory to be wasted,
because above 4K `absl::CordBuffer` imposes size classes being powers of 2 minus
13, which are generally more sparse than those of the memory allocator.

PiperOrigin-RevId: 607921546