Exempt "/proc/*/fd" files from the assumption that if they have zero length then they do not support random access.
In particular `CFile{Reader,Writer}` can easily have such a `filename()` when
they are created with an already open `FILE*`.

For these files we don’t really have any signal based on the filename. Most
likely they are regular files which support random access, or devices or pipes
which do not.

Cosmetics: in `{Fd,CFile}Reader`, move the bulk of random access detection to a
separate function, to code early return more easily and to reduce nesting,
consistently with `{Fd,CFile}Writer`.
PiperOrigin-RevId: 608508319