Change the behavior of `BackgroundCleaner::ScheduleCleaning()` to never extend a pending deadline.
This will allow scheduling cleaning after unlocking a mutex which guards
computing when to clean, without the danger of postponing another cleaning which
was scheduled in the meantime.

The cost is that the client might get calls to `Clean()` even if they are no
longer needed.

This actually simplifies the implementation, which can quickly rule out
extending the deadline, can start scanning from the current position instead of
from `entries_.end()` when reducing the deadline, and does not need to consider
encountering the old position of the given entry while looking for the right
place. OTOH all `entries_` before `next_` must be marked (here with
`deadline == absl::InfiniteFuture()` which is impossible for entries after
`next_`), so that it is possible to tell whether the given entry is before of
after `next_`.

PiperOrigin-RevId: 613114506