Shard `{,Keyed}RecyclingPool::global()` instances across threads by introducing `RecyclingPoolOptions::thread_shards()`.
`{,Keyed}RecyclingPool::global()` maintain this many shards of pools, with the
shard to use based on the thread which called `global()`.

A larger number of shards increases the probability that objects are recycled by
the same thread which has created them, and decreases lock contention, at the
cost of increasing memory usage.

This is not yet the default but will be soon.

PiperOrigin-RevId: 613158880