In `DependencyDefault` specialization where `DependencyManagerRef<Manager>` is used, specify `DependencyManagerRef<DependentManager>` instead of `DependencyManagerRef<Manager>` as the result of the `get()` overload which returns the original type.
This avoids a compile error when `DependencyManagerRef<Manager>` is an array
type. The overload is disabled with SFINAE in this case (given that `Handle` is
definitely not a supertype of an array type), but a function must not be
declared as returning an array even if SFINAE disables it.

Do the same for `DependencyManagerPtr<Manager>` for consistency, even though it
is currently never an array type.

PiperOrigin-RevId: 616819957