Fix adopting `AnyDependency` storage in case the target `AnyDependency` has `inline_size == 0` (which imposes additional constraints on the stored `Dependency` regarding its stability and trivial relocatability), while the source `AnyDependency` stores a `Dependency` which does not satisfy the additional constraints, yet its size fits.
In this case the storage used to be adopted, and the resulting `AnyDependency`
was declaring itself stable, even though the stored `Dependency` was not
necessarily stable.

Also, change `IsInline` from a type trait to a constexpr function. This will
simplify a future change when available size and alignment are known only at
runtime.

PiperOrigin-RevId: 622207722