Optimize moving or adopting `AnyDependency{,Ref}` when `inline_size == 0` by replacing an indirect call to `methods_and_handle_.methods->move()` with a plain assignment of `methods_and_handle_.handle` and `repr_`.
If `ABSL_ATTRIBUTE_TRIVIAL_ABI` is not available, let `inline_size == 0`
require `std::is_trivially_copyable` (a stronger condition) instead of
`absl::is_trivially_relocatable` (previously this constraint was skipped
in this case). This is relied upon in a new context which is not guarded
with `ABSL_ATTRIBUTE_TRIVIAL_ABI`.

PiperOrigin-RevId: 622825043