Add `InitializerTarget<T>::type` and `InitializerTargetT<T>` which deduce the appropriate target type such that `T` is convertible to `Initializer<InitializerTargetT<T>>`.
This allows a single template to uniformly handle a `Target` passed directly and
`Initializer<Target>`. This is also useful for CTAD guides to deduce a template
argument as `InitializerTargetT<T>`.

This is undefined in the case of `std::tuple<Args...>` which requires the target
type to be specified by the caller.

Use `InitializerTargetT` to simplify CTAD guides: there is no need to
explicitly exclude `std::tuple`. Make their comments more accurate: passing
`Initializer<T>` to a constructor deduces the template argument as `T`,
not as `Initializer<T>`.

Use `InitializerTargetT` to uniformly handle `Target` passed directly or as
`Initializer<Target>` in functions which deduce the type from an argument:
`MakeAnyText{Reader,Writer}()`.

PiperOrigin-RevId: 623446209