Introduce `riegeli::Maker()`, `riegeli::Maker<T>()`, and `riegeli::Maker<Template>()` to replace `std::forward_as_tuple()` for specifying constructor arguments for construction in-place.
Rationale. For a long time Riegeli has had two styles for passing dependent
objects:

1. Using CTAD:
```
riegeli::CsvWriter csv_writer(
    riegeli::FdWriter(filename),
    riegeli::CsvWriterBase::Options().set_header(*kHeader));
```

2. Using `std::forward_as_tuple(args...)` with constructor arguments for
construction in-place:
```
riegeli::CsvWriter<riegeli::FdWriter<>> csv_writer(
    std::forward_as_tuple(filename),
    riegeli::CsvWriterBase::Options().set_header(*kHeader));
```

The first is more readable, especially with deep nesting (avoids a separate
nesting of types and a parallel nesting of `std::forward_as_tuple()` calls),
even though it is prone to the most vexing parse issue. The second is more
efficient (avoids moving temporaries).

This introduces a new style which combines the benefits of both:

3. Using CTAD and `riegeli::Maker<T>(args...)` with constructor arguments for
construction in-place:
```
riegeli::CsvWriter csv_writer(
    riegeli::Maker<riegeli::FdWriter>(filename),
    riegeli::CsvWriterBase::Options().set_header(*kHeader));
```

Usages of `std::forward_as_tuple()` for this purpose will be phased out, and the
preferred style will be `riegeli::Maker<T>()`, or `riegeli::Maker()` if `T` is
apparent nearby (e.g. in `std::make_unique<Template<T>>`) or if CTAD cannot be
used (Riegeli code itself does not assume C++17 yet).

`riegeli::Maker<T>(args...)` provides a syntax for specifying constructor
arguments together with the type to be constructed — also for other contexts,
including `AnyDependency`, and functions which deduce the type from the argument
like `{Reader,Writer}CFile()`.

Also add `riegeli::OwningMaker()`, `riegeli::OwningMaker<T>()`, and
`riegeli::OwningMaker<Template>()` which owns the arguments.

The type of `riegeli::Maker(args...)` is `MakerType<Args...>`, and the type of
`riegeli::Maker<T>(args...)` and `riegeli::Maker<Template>(args...)` is
`MakerTypeFor<T, Args...>`.

Note that `riegeli::Maker<T>(args...)` cannot directly return `Initializer<T>`
for a subtle technical reason: `Initializer<T>` erases the type of `args`,
so it needs to pack all arguments in a single object to be referred as `void*`.
`riegeli::Maker<T>(args...)` would have no good way to allocate that object in
the context of the caller, because a variadic function cannot have a parameter
with a default value, which is the usual trick for such cases. Having two
stages, constructing `MakerType<T, Args...>` and converting it to
`Initializer<T>`, allows `Initializer` to refer to the `MakerType` as the
type-erased storage.

`RiegeliDependencySentinel(T*)` continues to return a value convertible to
`Initializer<T>`, but this change implies that it should return
`MakerType<Args...>` instead of `std::tuple<Args...>`.

Remove support for `std::in_place_type` and `riegeli::in_place_template` in
`AnyDependency`. It is more readable to use `riegeli::Maker<T>()`.

Remove `AnyDependency::Reset()` with a single parameter. This is the same as
assignment (which is defined for consistency with implicit conversions), and
nowadays `riegeli::Reset()` uses assignment for resetting to a single value even
when the type does not match but is convertible.

Deprecate `AnyDependency::Emplace<T>(args...)`. This is the same as assigning
`riegeli::Maker<T>(args...)` (except for the return value), which is more
readable than the old alternative of `Reset(std::in_place_t<T>, args...)`.

Remove support for constructing `AnyDependencyRef` from `Initializer`. That was
done for consistency, but that did not add new functionality, because passing a
temporary `Manager` to `AnyDependencyRef` is as efficient as constructing it in
place. If some code wants to construct `AnyDependencyRef` from an `Initializer`
for some reason, an explicit call to `Initializer::Reference()` can be used.

Simplify `AnyDependency` constructors: they all use a single implementation with
`InitializerTargetT<Manager>` as the stored `Manager` type.

Fix outdated comments: `AnyDependencyRef` no longer derives from
`AnyDependency`.

Change `Chain::FromExternal<T>(std::forward_as_tuple(args...))` to
`Chain::FromExternal(riegeli::Maker<T>(args...))`. This effectively removes
support for creating immovable objects before C++17 where copy elision was not
guaranteed.

Use `InitializerTargetT` to uniformly handle `Target` passed directly or as
`Initializer<Target>` in remaining functions which deduce the type from an
argument: `{Reader,Writer}CFile()`.
PiperOrigin-RevId: 624163409