Add `AnyDependencyInitializer`.
A parameter of type `AnyDependencyInitializer<Handle>` allows the caller to
specify an `AnyDependency<Handle>` by passing a value convertible to
`AnyDependency<Handle>`.

In contrast to accepting `AnyDependency<Handle>` directly, this allows to
construct the object in-place, avoiding constructing a temporary and moving
from it. This also avoids specifying `::Inlining<...>` in the interface while
benefiting from that in the implementation.

This is similar to `Initializer<AnyDependency<Handle>>`, except that it
efficiently handles `AnyDependency<Handle>` specializations with any inline
storage constraints.

`AnyDependencyInitializer<Handle>(manager)` does not own `manager`, even if it
involves temporaries, hence it should be used only as a parameter of a function
or constructor, so that the temporaries outlive its usage. Instead of returning
an `AnyDependencyInitializer<Handle>` from a function, consider returning
`MakerTypeFor<Manager, ManagerArgs...>` or `AnyDependency<Handle>`.

PiperOrigin-RevId: 624166601