Make move constructors and move assignment operators properly conditionally available when movability and/or assignability depends on type parameters.
This lets `std::is_move_{constructible,assignable}` yield correct results
for instantiations which lack some of these properties. This in turn allows
`Initializer` to compile for such types (type erasure in `Initializer` requires
assignability to be SFINAE-detectable even if `Initializer::AssignTo()` is not
actually called).

For example a lambda is move constructible but not move assignable.
Hence a `Digester` containing that lambda is by default not assignable,
and a `DigestingReader` containing that `Digester` should be SFINAE-detectable
as not assignable.

For some classes this is done by just making the move constructor and move
assignment defaulted. For other classes this uses a new
`MovingDependency<Handle, Manager, Mover>` class.

As a bonus, some move constructor no longer require the corresponding
`Dependency` to be default-constructible, and use its move constructor
instead of move assignment.

Consistently avoid accessing pointers to internals of moved-from objects (for
pointer arithmetic or comparing against `nullptr`). Even though this is harmless
in practice, formally this has UB, and there might be some sanitizers which
trigger on that.

Add assertions for invariants regarding buffer pointers being relied upon during
moving. The invariants allow to skip saving `start_to_limit()` across the moove
when it can be computed otherwise.

PiperOrigin-RevId: 626339916