Fix moving `{ReaderI,WriterO}Stream` and other classes using virtual inheritance on Windows.
The Windows C++ ABI supports conversion of member pointers involving virtual
inheritance. In order to do this, it uses a different representation of a member
pointer when virtual inheritance is involved, which accesses the virtual table.
This is incompatible with how `MovingDependency` finds the host object.

Since with virtual inheritance the move constructor and assignment generally
need to be defined explicitly anyway, `MovingDependency` is extended with
passing `this_host` and `that_host` explicitly.

This is done this way instead of reverting to performing work in the body of the
move constructor, because this allows injecting code and capturing state before
move-constructing the `Dependency`, without requiring the `Dependency` to be
default-constructible.

PiperOrigin-RevId: 632460112