Do not bother type-erasing `MemoryEstimator` support in `SharedPtr`. Store just the `destroy` method, removing an indirection through the method table.
This means that its memory estimation is about as accurate as for
`std::unique_ptr`: it relies on `DynamicSizeOf()` and `RegisterSubobjects()`
called on the base class. Also, the allocated pointer value and the accurate
header size are not used for estimation if the derived class is unknown, because
the alignment of the derived class may differ from the alignment of the base
class.

The benefits are reduced coupling between libraries and less code generated.
`SharedPtr` is meant to be more low level.

PiperOrigin-RevId: 632460330