Use more `IntrusiveSharedPtr<Chain::RawBlock>` instead of `Chain::RawBlock*`.
Let `{Set,Pop}{Front,Back}()` return the old block pointer as
`IntrusiveSharedPtr<Chain::RawBlock>`.

Let `Chain::RawBlock::Copy()` never drop ownership of the caller. This is now
handled when the caller gets removed from the block pointer array.

Change `Chain::{Append,Prepend}RawBlock()` to unconditionally take ownership.
This makes the API simpler. It is more common when they use the block anyway,
so only rarely this causes an extra `Ref()` and `Unref()` pair, and if the block
turns out to ont be rewritten, this is expensive anyway.

PiperOrigin-RevId: 639694350