Improve detection of "/proc", "/sys", "/dev/cgroup" etc. files which claim to have zero size but have non-empty contents when read.
Instead of trying to recognize them by filenames, which is brittle, if the
reported size is 0 and the current position is 0, read something to the buffer.
If that succeeds, assume that this is one of these unusual files, and disable
random access together with caching the size.

If the `growing_source` option is `true`, check the size again, because another
explanation is that the file could have grown, in which case random access is
fine (and the size is not cached in this case anyway).

This can now be applied also to `IStreamReader` which does not know about
filenames.

This should be reasonably cheap: for regular files it involves extra work only
for empty files, and trying to read an empty file does not need to bring file
contents.

Disable checking of the filename in `Writer` classes. It is not that important
because the size is not cached in this case, and these files are rarely written
to anyway.

Some "/sys" files claim to have the size of 4096. They are no longer excluded
from random access and caching the size, which is fine in practice as long as
real contents are not larger than the claimed size.

Cosmetics: when checking the size requires seeking to the end, avoid the syscall
of seeking back if the target position already matches the end position.
PiperOrigin-RevId: 642438458