Generalize `SharedPtr<T>::Reset()` optimization of assigning in place if the object is unique, from the argument of just `Initializer<T>` to other cases when the argument can be used to construct exactly `T`, e.g. `T` itself or `MakerTypeFor<T, Args...>`.
These cases fell into the overload for subtypes, which was a better match than
`Initializer<T>` because it does not need converting an argument.

A possible fix is to ensure that subtypes are proper subtypes. This avoids any
potential for ambiguity: exact matches of the type are handled by converting to
`Initializer<T>`.

Also, introduce a private type trait to make it obvious that `ResetImpl()`
overloads are mutually exclusive.

PiperOrigin-RevId: 642529998